Improve serializer resilience against "infinite" loops
The (de-)serializer now stores only one error and noops any operation
after that error is registered. This should protect from extremly long
Seq loops when a big number os accidentally deserialized as a Seq (or
Map) length. It also makes more sense to immediately abort and not
collect errors, because if one field was wrong, the rest of the data is
unlikely to be correctly (de-)serialized.

Also add a Finish() function to the deserializer which checks if the
input buffer has been completely consumed - this could also help to catch
mistakes.