Do only retry TCP requests on write comm related errors
Before the request would be retried regardless of the error. This
was wrong because other errors (e.g serialization) won't be fixed
by retrying. In some cases this could also cause harm: If the server
already sent a response and deserialization fails, retrying causes
double processing of the request - and BeeMsg is not guaranteed to be
idempotent.

Now we only retry if anything relating to writing the request to
the buffer fails (e.g. to a closed connection). We intentionally ignore
comm fails on reading the response as we must assume that the server
has already processed the request if it received the complete request.