Add (De)SerializeStringSeq to BeeSerde
This is a specialization that is used on the C++ side by
default for std::list<std::string>, std::vector<std::string> and
std::set<std::string>. It skips the length fields that CStr has and
totally relies on a terminating null byte for deserialization. For that
reason, the serialization input is checked for containing null bytes
which will cause an error and abort serialization (it would break the
deserializer).

On the deserializing side, we have to accept the fact that std::string can
potentially contain null bytes and break the deserializer when sent here.
The C++ code suffers from the same problem, but I never saw that actually
happening, so it should be fine.

Note that we refer to "strings" here despite accepting []byte because on the
C++ it is used for std::string, which is not the same as a string in Go.

Also remove some unnecessary type definitions in TestNestedSeq (the linter complains about that).