Cosmetic improvements to alerts and health output
Change the behavior of the terminal alert to only trigger the bell and
not flash the terminal manually as well. That way, whether the bell is
audible, visible or both is up to the user's terminal configuration.

Also, change the wording in the health output to only suggest running
fsck in readonly mode only.