Update path to copy binary, add threads flag, update docs
The num-workers flag is used to automatically determine the number of Goroutines CTL should use.
This might be confusing for the copy mode where threads often applies to other nodes, not just this one.
So it is better to have a dedicated flag for threads.
This also makes it more straightforward to customize the default and hold one thread back for CTL.

Incorporates the remaining man page docs into the copy help.