Fix remote pull and simplify usage
The added check for client sessions did not handle if the file did not already exist, which is usually the case for downloads.

When downloading files require the user always specify the remote path to clearly differentiate between the name of the file being downloaded, and the location it is being downloaded inside BeeGFS (which typically won't already exist).