Fix check when setting mirrored stripe pattern
This was incorrectly checking if there were available metadata buddy groups when it should check for storage buddy groups.

The original CTL only verified there was at least one storage buddy group created in the overall system.
This new implementation verifies there are buddy groups or targets available in the pool configured for this entry based on the requested stripe pattern.