By default require root privileges to execute all commands
Decouples frontend checking of EUIDs to determine if the user can execute commands from backend functionality that uses the EUID for other purposes.

Fixes a bug in the entry package where an error walking directories would cause a panic (as can now happen if a user calls entry info recursively and encounters entries they do not have permission to access).