Address review
Also change flags to --uids and --gids to be in line with set-limits.
Also change default behavior to apply only if neither of the above are
set to anything. Otherwise, if one gave e.g. --uids=all, it would still
include the current group ids, which is probably not wanted.