Ensure stack traces can be printed when the euid is root
If the CTL binary was installed using a package it will have the setgid bit set. This causes Go programs to act as though GOTRACEBACK=none is set and stack traces are suppressed for security reasons.

Because this makes debugging panics challenging, if the user has root privileges anyway, reset GOTRACEBACK back to the default single configuration or print extra text explaining the lack of traces.

Reference:
* https://github.com/golang/go/issues/62474#issuecomment-1708447731
* https://github.com/golang/go/issues/60272