filesystem: optionally walk directory entries in lexicographical order
Added a new WalkDirLexicographically function that is identical to the built-in filepath.WalkDir except it walks paths lexicographically so the order is the same as iterating over paths in BadgerDB.
Note if BadgerDB contains entries for directories they MUST end in a slash for the sorting to be consistent.

By default BeeGFS.WalkDir will not sort lexicographically and callers can enable this as needed using new walk options.