ctl: make remote push/pull commands idempotent
Add support when Remote returns an existing job with a response status.

Extend the force flag so it can also be used to force job creation even if a path is already in sync.

Improve how job results are summarized bringing them more in line with the new remote status command.