feat: Make `stats client|user` compatible with ipv6 server nodes
The ipv6 capable servers can report ipv6 addresses, which is only
possible
with a new GetClientStatsV2 message. To maintain compatiblity with not
updated legacy nodes, this includes a mechanism to detect wether
or not use the new V2 message or the old one, so this should work
with a mixed system of pre- and post ipv6 nodes.