fix(ctl): path handling with unmounted filesystems
Only the walkDir (recurse) mode requires a mounted file system. The other modes walkStdin and
walkList can be used when BeeGFS is unmounted, unless a filter is set. Restructure the function so
initializing the client happens first, and ErrUnmounted is ignored. If a filter is applied it now
uses the client.Lstat method which returns a specific error if BeeGFS is unmounted.